Fix pointer truncation in au4MakeDitherTable for x64 (x64 runtime fix)
CLightBlend::au4MakeDitherTable stored the dither buffer address in a
uint32 (`mem_addr = (uint32)au4AlphaRefDitherBuf`), did alignment math on
it, then cast it back to a pointer (`table_ptr = (uint32*)mem_addr`). On
x64 the (uint32) cast truncates the 64-bit heap pointer to 32 bits, so
table_ptr is garbage and the table fill writes through it -> access
violation at startup (crash observed on the first x64 run,
LightBlend.cpp:485). Change mem_addr to uintptr and use (uintptr) casts;
on Win32 uintptr==uint32 so behaviour is identical. This is an LLP64 bug
the compile-driven hygiene sweep missed because (uint32)ptr is only a
C4311 warning, not an error.

+++ jp2_pc/Source/Lib/Renderer/LightBlend.cpp
@@ -385,7 +385,7 @@ uint32 *CLightBlend::au4MakeDitherTable
 //
 //**************************************
 {
-	uint32 mem_addr;
+	uintptr mem_addr;
 	uint32 *table_ptr;
 
 #if (DITHER_SIZE == 2)
@@ -402,7 +402,7 @@ uint32 *CLightBlend::au4MakeDitherTable
 
 	AlwaysAssert(au4AlphaRefDitherBuf);
 
-	mem_addr = (uint32)au4AlphaRefDitherBuf;
+	mem_addr = (uintptr)au4AlphaRefDitherBuf;
 
 	//
 	// Insure alignment on a table sized boundry.
@@ -451,7 +451,7 @@ uint32 *CLightBlend::au4MakeDitherTable
 
 	AlwaysAssert(au4AlphaRefDitherBuf);
 
-	mem_addr = (uint32)au4AlphaRefDitherBuf;
+	mem_addr = (uintptr)au4AlphaRefDitherBuf;
 
 	//
 	// Insure alignment on an 4*width*shades*entries boundry.
